backend/requirements.txt
# Core Agent Framework
langgraph>=1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# Vector Store & Embeddings
chromadb>=0.4.0
langchain-google-genai>=0.0.1
sentence-transformers>=2.2.0

# Web Framework (Phase 2)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database (Phase 2)
motor>=3.3.0
pymongo>=4.6.0

# Search APIs
tavily-python>=0.2.0
arxiv>=2.1.0

# PDF & Document Processing
pymupdf>=1.23.0
python-docx>=0.8.11
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
